updated graphql api
